LayoutTests/webcl/tests/webcl-test-arg-types-const-expected.txt

WebCL WebCLKernelArgTypes constants tests.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test all constants defined in WebCLKernelArgTypes.idl
PASS webcl.CHAR == 0 is true
PASS webcl.UCHAR == 1 is true
PASS webcl.SHORT == 2 is true
PASS webcl.USHORT == 3 is true
PASS webcl.INT == 4 is true
PASS webcl.UINT == 5 is true
PASS webcl.LONG == 6 is true
PASS webcl.ULONG == 7 is true
PASS webcl.FLOAT == 8 is true
PASS webcl.HALF == 9 is true
PASS webcl.DOUBLE == 10 is true
PASS webcl.VEC2 == 0x0100 is true
PASS webcl.VEC3 == 0x0200 is true
PASS webcl.VEC4 == 0x0400 is true
PASS webcl.VEC8 == 0x0800 is true
PASS webcl.VEC16 == 0x1000 is true
PASS webcl.LOCAL_MEMORY_SIZE == 255 is true
PASS successfullyParsed is true

TEST COMPLETE
